// src/tcgcommerce/modules/import/sort/data/dto/import.sort.data.dto.ts
export class ImportSortDTO {
    importJobId: string;
    importSortTotalCardQty: number;
    importSortTotalCardTCGPlayerMarketPrice: number;
    importSortTotalCardTCGPlayerLowPrice: number;
    importSortCards: ImportSortCardDTO[];
}

export class ImportSortCardDTO {
    importSortCardTCGPlayerId: number;
    importSortCardName: string;
    importSortCardSetName: string;
    importSortCardNumber: string;
    importSortCardCondition: string;
    importSortCardPrinting: string;
    importSortCardTCGPlayerMarketPrice: number;
    importSortCardTCGPlayerLowPrice: number;
    importSortCardQty: number;
}